Optimize for turn in now

+++ Structs.h
@@ -1,5 +1,5 @@
 typedef struct obj_camera{
-    float width, height;
+    double width, height;
 } obj_camera;
 
 typedef struct obj_sphere{
@@ -16,13 +16,13 @@ typedef struct obj_plane{
 
 typedef struct scene_object{
         char type;
-        float *color;
+        double *color;
         //What is position relative to?
         //Position is relative from origin in 3D space
         //Camera is at origin and looking down positive z
-        float *position;
-        float *normal;
-        float radius;
+        double *position;
+        double *normal;
+        double radius;
 
 } scene_object;
 typedef struct pixels{
